test(schema): move integration tests
