+++ src/lib.rs
@@ -28,5 +28,5 @@ pub mod value;
 
 pub use driver::Driver;
 pub use error::Error;
-pub use schema::{Definition, Type, Schema};
+pub use schema::{Definition, Type, DriverConfig, Schema};
 pub use value::{Key, Pointer, Object, Array, Value};
